Refactor code for improved readability and dependency imports
Reorganized the properties in the DsButton, DsAlert, and DsLink components for better readability. This makes it easier to recognize the property associations. The imports of DsButton, DsTypography, DsLink and DsAlert have also been updated to use local reference instead of "@perezguedesmaikel/chile-atiende-ui". This should reduce potential dependency issues moving forward.